@@ -41,23 +41,23 @@
   },
   "config": {},
   "peerDependencies": {
-    "@angular/common": ">=4.3.0",
-    "@angular/core": ">=4.3.0",
-    "@ngx-translate/core": ">=7.0.0"
+    "@angular/core": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@ngx-translate/core": "^10.0.0"
   },
   "dependencies": {
     "ltx": "^2.7.1"
   },
   "devDependencies": {
-    "@angular/core": "^4.3.0",
-    "@angular/common": "^4.3.0",
-    "@angular/platform-browser": "^4.3.0",
-    "rxjs": "^5.1.0",
-    "zone.js": "^0.8.4",
-    "typescript": "^2.2.2",
-    "@ngx-translate/core": "^7.0.0",
-    "tslint": "^4.5.1",
-    "tslint-eslint-rules": "^3.4.0",
+    "@angular/core": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "rxjs": "^6.2.2",
+    "zone.js": "0.8.26",
+    "typescript": "2.7.2",
+    "@ngx-translate/core": "^10.0.0",
+    "tslint": "5.8.0",
+    "tslint-eslint-rules": "^4.1.1",
     "rimraf": "^2.6.1"
   }
 }
